refactor: move skeleton + component extraction into API, CLI is now 69 lines
All template logic (list, copy, skeleton, component extraction) lives
in api/template.mjs. CLI handler is just arg parsing + text formatting.

Made-with: Cursor
